backend/requirements.txt
easyocr
opencv-python
numpy
pillow
pdf2image